86cym3xfh: Refactor components in DraftListingPage to seperate files
